fix: JpaRepository PK 타입 수정

+++ src/main/java/com/itwillbs/factron/repository/syscode/DetailSysCodeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetailSysCodeRepository extends JpaRepository<DetailSysCode, Integer> {
+public interface DetailSysCodeRepository extends JpaRepository<DetailSysCode, Long> {
 }
